Split core code into `core` and `backend` packages
