+++ core/jvm/src/main/scala/eu/joaocosta/minart/backend/JavaRenderLoop.scala
@@ -1,13 +1,11 @@
-package eu.joaocosta.minart
+package eu.joaocosta.minart.backend
 
 import scala.annotation.tailrec
 import scala.concurrent.duration.FiniteDuration
-import scalanative.native._
 
-import sdl2.SDL._
-import sdl2.Extras._
+import eu.joaocosta.minart.core._
 
-object SdlRenderLoop extends RenderLoop {
+object JavaRenderLoop extends RenderLoop {
   def finiteRenderLoop[S](
     canvasManager: CanvasManager,
     initialState: S,
@@ -24,8 +22,8 @@ object SdlRenderLoop extends RenderLoop {
       val startTime = System.currentTimeMillis()
       val newState = renderFrame(canvas, state)
       val endTime = System.currentTimeMillis()
-      val waitTime = scala.math.max(0, frameMillis - (endTime - startTime))
-      if (waitTime > 0) SDL_Delay(waitTime.toUInt)
+      val waitTime = math.max(0, frameMillis - (endTime - startTime))
+      if (waitTime > 0) Thread.sleep(waitTime)
       if (terminateWhen(newState) || !canvasManager.isCreated()) ()
       else finiteRenderLoopAux(newState)
     }
@@ -38,13 +36,5 @@ object SdlRenderLoop extends RenderLoop {
     renderFrame: Canvas => Unit): Unit = {
     val canvas = canvasManager.init()
     renderFrame(canvas)
-    var quit = false
-    val event = stackalloc[SDL_Event]
-    while (!quit) {
-      while (SDL_PollEvent(event) != 0) {
-        if (event.type_ == SDL_QUIT) { quit = true }
-      }
-    }
-    canvasManager.destroy()
   }
 }
