+++ core/shared/src/main/scala/eu/joaocosta/minart/core/CanvasManager.scala
@@ -1,4 +1,4 @@
-package eu.joaocosta.minart
+package eu.joaocosta.minart.core
 
 /**
  * Abstraction that provides `init` and `destroy` operations to
