+++ core/shared/src/main/scala/eu/joaocosta/minart/core/Color.scala
@@ -1,4 +1,4 @@
-package eu.joaocosta.minart
+package eu.joaocosta.minart.core
 
 /** RGB Color */
 final case class Color(r: Int, g: Int, b: Int)
