+++ core/shared/src/main/scala/eu/joaocosta/minart/core/FrameRate.scala
@@ -1,4 +1,4 @@
-package eu.joaocosta.minart
+package eu.joaocosta.minart.core
 
 import scala.concurrent.duration._
