+++ examples/js/src/main/scala/eu/joaocosta/minart/examples/JsColorSquare.scala
@@ -1,6 +1,6 @@
 package eu.joaocosta.minart.examples
 
-import eu.joaocosta.minart._
+import eu.joaocosta.minart.backend._
 
 object JsColorSquare {
   def main(args: Array[String]): Unit = {
