+++ examples/jvm/src/main/scala/eu/joaocosta/minart/examples/JvmFire.scala
@@ -1,6 +1,6 @@
 package eu.joaocosta.minart.examples
 
-import eu.joaocosta.minart._
+import eu.joaocosta.minart.backend._
 
 object JvmFire {
   def main(args: Array[String]): Unit = {
