+++ pure/shared/src/main/scala/eu/joaocosta/minart/pure/MinartApp.scala
@@ -1,6 +1,6 @@
 package eu.joaocosta.minart.pure
 
-import eu.joaocosta.minart._
+import eu.joaocosta.minart.core._
 
 trait MinartApp {
   type State
